remove C++ API wrappers and tests
