+++ node.gyp
@@ -241,7 +241,6 @@
       'src/node_dir.h',
       'src/node_dotenv.h',
       'src/node_embedding_api.h',
-      'src/node_embedding_api_cpp.h',
       'src/node_errors.h',
       'src/node_exit_code.h',
       'src/node_external_reference.h',
@@ -1278,14 +1277,6 @@
         'test/embedding/embedtest_c_api_run_main.c',
         'test/embedding/embedtest_c_api_threading.c',
         'test/embedding/embedtest_c_api.c',
-        'test/embedding/embedtest_c_cpp_api_common.cc',
-        'test/embedding/embedtest_c_cpp_api_common.h',
-        'test/embedding/embedtest_c_cpp_api_env.cc',
-        'test/embedding/embedtest_c_cpp_api_modules.cc',
-        'test/embedding/embedtest_c_cpp_api_preload.cc',
-        'test/embedding/embedtest_c_cpp_api_run_main.cc',
-        'test/embedding/embedtest_c_cpp_api_threading.cc',
-        'test/embedding/embedtest_c_cpp_api.cc',
         'test/embedding/embedtest_main.cc',
       ],
